move memory settings extraction to cli, also use it for skip compilation logic
